+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 7.81
 -----
+*   New Features
+    *   Add support for removing individual episodes from listening history.
+        ([#3453](https://github.com/Automattic/pocket-casts-android/pull/3453))
 *   Bug Fixes
     *   Fix the cropping issue with the Add 5 Minutes sleep timer button.
         ([#3436](https://github.com/Automattic/pocket-casts-android/pull/3436))